requirements.txt
pdfplumber>=0.11.0
langchain-text-splitters>=0.3.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
sentence-transformers>=3.0.0
chromadb>=0.5.0
python-dotenv>=1.0.0
